Combined both spot and futures
